backend/products/serializers/tag.py
from rest_framework import serializers

from products.models.tag import Tag


class TagSerializer(serializers.ModelSerializer):
    class Meta:
        model = Tag
        fields = ('id', 'name', 'slug')
