backend/requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.2
django-cart==1.0.4
django-cors-headers==4.2.0
django-environ==0.10.0
django-extensions==3.2.3
django-filter==23.4
django-phonenumber-field==7.2.0
django-redis==5.2.0
django-sendsms==0.2.3
django-smsru==1.0.4
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-extra-fields==3.7.0
drf-spectacular==0.26.5
filetype==1.2.0
flake8==6.1.0
flake8-bugbear==23.9.16
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Markdown==3.0
mccabe==0.7.0
oauthlib==3.2.2
pep8-naming==0.13.3
phonenumbers==8.13.23
phonenumberslite==8.13.23
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.3
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
